Use Exceptions to simplify input error handling
Inside GameCommand.cpp->CommandHandling(), checks if the input is a
valid command code and also checks if the parameters are integers.
Replaced relevant cout statements in Do_Command() functions with a
throw to the catch block. Still need to keep the cin.clear() at the
start of try block to prevent infinite loops.